MailTracker/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
